reload if the file changes; display error in status on error
This command is useful when we want to apply some sort of automatic
formatting tool on save. In that case, if the file is changed, we want
to reload it. Use the file's mtime to detect this case.

Additionally, if the command fails, we want to display a message. Show
the error in the status bar, as well as logging to the console.